+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 # 1.5.3
 - Geometry grid is now triplanar
-- Improve edge selection highlighted when grid view is on
+- Improve z fighting on grid overlay
+- Improve edge selection highlighted when grid overlay is on
 
 # 1.5.2
 - Allow users to customise combine face angle limit (hotspotting)